requirements.txt
requests==2.26.0
beautifulsoup4==4.9.3
selenium==4.0.0.b4
pandas==1.3.1
python-dotenv==0.19.0